requirements.txt
# Python 3.10.16
numpy>=1.24.3
pandas>=2.2.3
scikit-learn>=1.6.1
scipy>=1.15.3
matplotlib>=3.10.0
colour-science>=0.4.6
optuna>=4.4.0
scikit-image>=0.25.2
